# pkg/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Сервис для управления баннерами",
        "title": "Banner Service",
        "contact": {
            "name": "Ilya Lyakhov",
            "email": "ilja.sarasti@mail.ru"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/banner/{id}": {
            "delete": {
                "security": [
                    {
                        "AdminToken": []
                    }
                ],
                "description": "API layer function which process the request and deletes banner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Banners"
                ],
                "summary": "Deletes banner by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Banner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Banner successfully deleted"
                    },
                    "400": {
                        "description": "Incorrect provided data",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "401": {
                        "description": "User not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "403": {
                        "description": "User insufficient rights",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "404": {
                        "description": "Banner not found",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AdminToken": []
                    }
                ],
                "description": "API layer function which process the request and updates banner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Banners"
                ],
                "summary": "Updates banner by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Banner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Banner update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateBanner"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Banner successfully updated"
                    },
                    "400": {
                        "description": "Incorrect provided data",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "401": {
                        "description": "User not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "403": {
                        "description": "User insufficient rights",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "404": {
                        "description": "Banner not found",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "409": {
                        "description": "Banner with provided params already exists",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateBanner": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "object"
                },
                "feature_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminToken": {
            "description": "Admin access token",
            "type": "apiKey",
            "name": "token",
            "in": "header"
        },
        "UserToken": {
            "description": "User access token",
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}